+++ src/components/Button/Button.tsx
@@ -10,61 +10,59 @@ type ContentProps = {
   children?: any;
 };
 
-export type ButtonProps = ButtonHTMLAttributes<HTMLButtonElement> & {
-  buttonModifiers?: Array<string>;
-  sizeType?: Size;
-  fullWidth?: boolean;
-  fullHeight?: boolean;
-};
-
-export type Props = ButtonProps & ContentProps;
+export type Props = ButtonHTMLAttributes<HTMLButtonElement> &
+  ContentProps & {
+    buttonModifiers?: Array<string>;
+    size?: Size;
+    className?: string;
+    fullWidth?: boolean;
+    fullHeight?: boolean;
+  };
 
 const ContentHolder = styled.span`
   padding: 0.2rem;
-  display: grid;
-  grid-auto-flow: column;
-  grid-gap: 0.2em;
 `;
 
 const Content = ({ icon = null, iconPosition = 'left', children }: ContentProps) => {
   if (iconPosition === 'right') {
     return (
-      <ContentHolder>
-        {children}
+      <>
+        {children && <ContentHolder>{children}</ContentHolder>}
         {icon}
-      </ContentHolder>
+      </>
     );
   }
 
   return (
-    <ContentHolder>
+    <>
       {icon}
-      {children}
-    </ContentHolder>
+      {children && <ContentHolder>{children}</ContentHolder>}
+    </>
   );
 };
 
 const Button = ({
-  sizeType = 'md',
+  size = 'md',
   buttonModifiers,
+  className = '',
   icon,
   iconPosition,
   children,
   fullWidth,
   fullHeight,
-  disabled,
   ...rest
 }: Props) => (
   <ButtonWrapper
-    sizeType={sizeType}
+    className={className}
+    size={size}
     modifiers={buttonModifiers}
     fullWidth={fullWidth}
     fullHeight={fullHeight}
-    disabled={disabled}
     {...rest}
   >
     <Content icon={icon} iconPosition={iconPosition}>
-      {children}
+      {' '}
+      {children}{' '}
     </Content>
   </ButtonWrapper>
 );